Fixed `LifeVariation` being improperly assigned tue to wrong datatype

+++ Entities/HDFirearm.cpp
@@ -847,7 +847,7 @@ void HDFirearm::Update()
                     pParticle->SetRotAngle(particleVel.GetAbsRadAngle() + (m_HFlipped ? -c_PI : 0));
 					pParticle->SetHFlipped(m_HFlipped);
 					if (lifeVariation != 0 && pParticle->GetLifetime() != 0) {
-						pParticle->SetLifetime(std::max(static_cast<int>(pParticle->GetLifetime() * (1.0F + (particleCountMax > 1 ? lifeVariation - (lifeVariation * 2.0F * (pRound->ParticleCount() / particleCountMax)) : lifeVariation * RandomNormalNum()))), 1));
+						pParticle->SetLifetime(std::max(static_cast<int>(pParticle->GetLifetime() * (1.0F + (particleCountMax > 1 ? lifeVariation - (lifeVariation * 2.0F * (static_cast<float>(pRound->ParticleCount()) / static_cast<float>(particleCountMax))) : lifeVariation * RandomNormalNum()))), 1));
 					}
                     // F = m * a
                     totalFireForce += pParticle->GetMass() * pParticle->GetVel().GetMagnitude();
